+++ apps/decodex/src/state/store/leases/read/basic.rs
@@ -1,0 +1,31 @@
+use crate::{
+	prelude::Result,
+	state::{IssueLease, StateStore},
+};
+
+impl StateStore {
+	/// Read the run lease for one issue.
+	pub fn lease_for_issue(&self, issue_id: &str) -> Result<Option<IssueLease>> {
+		let state = self.lock()?;
+
+		Ok(state.leases.get(issue_id).cloned())
+	}
+
+	/// List all run leases.
+	pub fn list_leases(&self, project_id: &str) -> Result<Vec<IssueLease>> {
+		let mut state = self.lock_without_refresh()?;
+
+		self.refresh_project_run_metadata_state_locked(&mut state, project_id)?;
+
+		let mut leases = state
+			.leases
+			.values()
+			.filter(|lease| lease.project_id == project_id)
+			.cloned()
+			.collect::<Vec<_>>();
+
+		leases.sort_by(|left, right| left.issue_id.cmp(&right.issue_id));
+
+		Ok(leases)
+	}
+}
